Allow configuring parsing method for data files
When configuring a product, the files are expected to be json files and
thus are directly serialized as such. Since we now want to be able to
load files that are in a binary format, we need to make it possible to
configure this behavior. The load step will thus be separated into more
steps that can then be overwritten by a product to alter each phase,
such as the parsing phase.