Include levitation, u, and v variable names in location hash
These attributes may not always be the same (e.g. 'lev' vs 'plev') and
thus need to be configurable. The most straight forward way for this is
to include them in the options part of the location hash. When they are
not specified, default values will be used. They also trigger a rebuild
to allow easily fixing a possibly invalid variable name.