Allow specifying a file in the attributes in the URL
We need to be able to set the referenced filename somewhere and because
the filenames will -probably- not match a given pattern, this needs to
accept a complete filename. This retains the limitation of only allowing
files to be used that are under the `data/weather/current` directory
because the parsing would need to be extended to allow `/` slashes to
handle remote files. In the future, we might want to completely change
the hash structure to handle such a case (and remove other things we
don't need).