Allow dropping files to load instead of specifying local file
Initially we just added a new parameter to allow specifying a name that the
client will try to load from the server. This is however obviously limited since
 the server needs to have the file stored and available. In general, this will
not be the case and it's more likely that the user has the file locally or in
some network file share that they're connected to. Thus, allowing the user to
just drop in a file makes this process a lot simpler and allows for more use
cases.