Fix error when loading a file with invalid indices
The URL keeps the current time and height value the user has selected. When the
user loads a new file, these values will be carried over and tried to be used.
However, if they did not exist, for example file only has time index 0 but index
 1 was presented, this caused an error on load. This now resets the values if
they're out of range.