Allow specifying full url as file
We previously allowed specifying the name of a file that resided on the server
to be loaded. This was because the url used `/` to separate option values. Thus,
 using a full url would conflict with this as a url by default contains multiple
 forward slashes. Now, the hash uses `|` as separators which should not occur in
 normal URLs, as they should be encoded.

For now, this keeps the loading of files from the local server available, but
this will probably be removed as it has little use.