Do not ignore hash changes
Upon initially saving the internal model configuration, we persist it in the
hash of the url and marking the next update to be ignored. This is because the
browser will most likely fire a hashchange event which would contain the exact
same hash that we just set. However, if, for some reason, the hash stayed the
same, the event would not be called and we would thus not reset the ignore
state. If the user then changes the URL manually, e.g. to change the file, this
would be ignored.

We don't really have that much in the url anymore so any changes should be
parsed easily and should also not trigger any internal updates since the values
(should) stay the same.