Fix switching between local & remote file
When loading a local file (via drag'n'drop) we cannot save that into the URL, as
 we could not retrieve the file by name nor would it help when the link was
shared. However, when we had a remote file and now dropped in a local file, we
still need to clear the previous state.

We couldn't quite differentiate between a local and remote file so we now keep
this info. This info may come in handy later as well.