Allow for arbitrary overlays
We previously only added support for wind and temperature overlays and tried to
match those to the variables given. This will stay and additionally any other
variables that have the correct dimensions will be provided with a generic
overlay. Generic here is a color scale from the detected min value to the
detected max value and thus may not match an expected scale. Additionally, since
 these values could get quite small, we try to estimate the necessary precision
of the displayed values to allow comparison by the user.