Invert height slider for certain units
The slider currently represents the highest point on the left and the lowest on
the right. This works for pressure given that the lowest pressure (i.e. the
first value in the data) is also the highest point. However, if instead the file
 used meters for height, we'd need to turn this slider around as the lowest
meter value would denote surface, the lowest point. For now, we turn around the
scale for meter and kilometer, but this could be foiled if the datafile already
has the values inverted, e.g. starting with the highest meter value.