Better handling of different height value layouts
Recently we moved the check for which direction the height values will go by
checking if the dimension was tagged with the positive going down. While this
did help with distinguishing between m and Pa unit directions, it did not help
with the values themselves being out of order. When the values for an altitude
dimension with meter as unit started at the highest value, the display was
switched again.

This now takes in account both of these stats (inversion because of unit,
different order in data) to properly create a display that's representative.