First pass on irregular grids
This adds support for visualizing irregular grids. Irregular grids differ to the
 regular grids that we had so far in that they are composed of triangles that
may be shifted a bit in each direction. Thus each triangle now has index for its
 center with the corresponding lat & lon values.

It would be a lot of work to add an irregular grid display to the visualization,
 so instead we convert the irregular grid into a regular one that we then
display as usual. This can lead to empty spots where a triangle does not cover a
 grid point as well as grid points being covered twice. For now, this will
suffice, but will be addressed at a later stage.

Fixes climate-v/netcdf-rust-emscripten#9.