Workaround for particles flying through center
Sometimes, since after we upgraded d3, particles go towards the center and very
high velocities and thus are like small flashes. This appears to happen because
they reach the backside of the projection and would get wrapped around because
of the distortion. Even though clipping should handle such a case, it sometimes
does not.

This workaround now stops rendering if the x or y difference of the particle are
 too large to make sense (in this case 20). Such a value should make sure that
even fast winds will be kept.