Average grid from equator outwards
When averaging the grid from the equator outwards, instead of the arctic, we
have a much better chance at actually filling all missing parts, because of the
propagating nature of this averaging. However, when complete rows are missing,
it will still fail as currently at least two more defined points are required.