Use scientific notation for very small values
Previously, we tried to figure out how many digits after the decimal point we
need to display to allow the user to see the differences in the values. This
only really worked for when the largest values were also still small as it was
based in the difference between min and max.

To be more flexible and also handle all cases well, this now uses the scientific
 notations for any displayed values that are smaller than 0.001.