Dial back fix for particles that move too fast
Previously a fix was added for particles that got distorted really bad causing
them to fly fast and through the center. In the fix, we ignored any draw calls
for particles that moved "too much" for a single step. However, the amount what
was deemed too much as it can be reached by some very high speed winds
(>350km/h).  As such, the filter has been adjusted to let a bit more faster
moving particles through.