Use symlog scale instead of normal log for generic overlays
The logarithmic scale does not work if the values cross zero or for zero itself.
 For wind and temperature overlays, this will/should not happen as the wind
speed will never be negative (because then it would just be positive wind in
another direction) as well as the temperature (we cannot go below 0K). However,
the generic overlays might very well go below zero and to make sure this is
working, the symlog scale is used instead.