Fix error when clicking between color scale names
When clicking between the color scales, a click was still registered and set to
an undefined value causing an error. This was because the event handler got
accidentally registered to broadly including the container the colorscale names
were in. Or alternatively, the container shouldn't be a `span` and thus wouldn't
 get registered the event either.