Allow for setting custom bounds in scale
The bounds for the color scale were currently always provided by the overlay
that was displayed. It used the min and max values that were found in the data.
In many cases, however, the max (or min) found in the data is far from what
would be considered "high" (e.g. found in data: 10C, expected high: 60C). We
have taken this route because we really don't know what sensible values for
these could be.

By allowing the user to set custom bounds, we move the responsibility to them to
 set the bounds right for their use, as they should have more knowledge on what
is too large or too small. We still provide the default range with the values we
 were given by the data.

To achieve this, the scale UI had to be refactored a bit to collect the
functionality at a single location for better integration of the individual
components.