Use new wasm backend based on head-scratcher
This replaces the previous backend based on emscripten and netcdf-rs with
wasm-pack and head-scratcher. The main difference is that head-scratcher allows
us to only load a small part of data into the browser and does not require the
full file. As a result, we can now handle files that would normally not fit into
 the RAM of the user. We currently do not cache the data, which could cause
re-loading of previous data, but since we cannot estimate how much data can be
kept in the cache, have no way of knowing if we have space for cache or not.

A downside of this new approach is that it has to be run in a web worker as all
the internal read operations are expected to be done synchronously. However, the
 web generally only provides asynchronous (in one way or another) abstractions
with no way of converting them to synchronous, as it would block the one, and
only, thread. In a web worker, fortunately, the browser does provide a
synchronous abstractions for file reading and resource requests. This obviously
does not eliminate asynchronicity and just moves the boundary to the interface
between the web page and the worker, but it's easier to work with async on the
page.

This major change resulted in some smaller, less related changes to accommodate
the new interface. We no longer just use the variable names in many instances,
as it is much simpler now to just return a full structure, which includes all
the information we could need (name, attributes, type). We also load the
lat/long values just once to avoid requesting the same values over and over
again.