Forward errors from worker to api caller
When an error happens in the worker (e.g. wrong data length), the error was so
far swallowed at the boundary between the two. This could cause multiple issues
in the client as it would keep waiting for a result. This now catches the error
at the worker and forwards it as a proper response which the client can then
take action on. In most cases, this will cancel the current action and just
display the error - which is already more than it did before.