Handle variables with missing time and/or height dimension
So far, we only selected variables that had all the dimensions we expected,
which was time, height, and ncells/lat & lon. This made it easy in the beginning
 to handle the api access via wasm, but seems like an unnecessary limitation.
This now lifts that limitation by handling these variables as well and
reflecting in the UI that they do not have any time and/or height attached to
them. Note that this applies to variables that do not have the dimension
assigned to them, but it does not include variables that have a separate
dimension of length one assigned instead.