Fix possible overlap of worker requests
When an overlay is used (such as temperature) and the file has wind variables
present, it is possible that worker requests were overlapping and causing wrong
results. This was because the overlays were created in parallel, which may
result in the requests for the variables of both overlays to be requested around
 the same time and thus both listeners getting the same data. Responses from the
 worker were only separated by the kind of request, so requests of the same kind
 could not be differentiated, which was the case here.

This introduces a request ID that is more or less unique given the amount of
worker requests we make and can therefore now be properly identified even among
requests of the same kind.