Sort overlays by name
The variables we are given usually are unsorted and thus looking for one in the
UI may be a bit cumbersome. This may especially be the case if two files have
the same variables, but because of their layout they were shown in a different
order. This now always sorts the overlays (apart from wind + temperature)
according to their names.