Provide conversion functions for generic overlays for known units
This is a first step into providing more useful conversions for units that we
directly know, as we do for temperature and wind. Currently, we just show the
unit that was provided by the variable responsible for an overlay without any
conversions to other, related units. While we cannot provide conversions for all
 kinds of units, this adds initial support for some simple units, specifically
temperature. However, this could be extended to other units that make sense.