Reset status message manually after grid and renderer
We are displaying messages to the user to show them what is currently being
worked on. When displaying a message, the message will stay displayed until
something else overwrites the message. This used to be fine because the
application would always run from the first to last agent and the last would
clear the message resulting in no message displayed to the user if everything
was working properly. In our case, however, we may not reach the last agents for
 two reasons:
- No file to display
- No overlay selected

This causes the message of the agent that just ran before them to stick, even
though they've already finished running. Since we know they've finished
ordinarily we will just reset the message manually in the two cases once the
previous agents have finished.