Let head-scratcher decide if file is valid
So far we specifically checked for the file ending to match `.nc` in order to
determine if we could handle the file or not. This is not ideal because file
extensions don't actually say anything about the content, can differ or be
changed easily. Additionally, remote files were not checked at all which could
then still cause an error that was not accounted for.

The head-scratcher already bails on invalid files, which we can use to properly
deal with this issue. This does require a change in the head-scratcher as it
currently panics instead of returning a proper error we can react to. With that
change, we can handle this like any other javascript exception.